Extract the methods for generating a list of clusters
These will be used in the `init`
